Extract 'replace' library code into the 'chapel' module
